# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CUPHY EduTech Platform API for Calcutta University Physics Honours Students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      parameters:
        - name: batchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectListResponse"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create subject (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject (admin)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject (admin)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # LECTURES
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List tests
      parameters:
        - name: batchId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [test, quiz]
      responses:
        "200":
          description: Tests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestListResponse"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create test (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
  /tests/{testId}:
    get:
      operationId: getTestById
      tags: [tests]
      summary: Get test with questions
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestDetail"
    put:
      operationId: updateTest
      tags: [tests]
      summary: Update test (admin)
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    delete:
      operationId: deleteTest
      tags: [tests]
      summary: Delete test (admin)
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tests/{testId}/submit:
    post:
      operationId: submitTest
      tags: [tests]
      summary: Submit test answers
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTestBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
  /tests/{testId}/results:
    get:
      operationId: getTestResults
      tags: [tests]
      summary: Get test results for a test (admin)
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResultListResponse"
...
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      parameters:
        - name: batchId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [notice, offer]
      responses:
        "200":
          description: Notices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeListResponse"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create notice (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{noticeId}:
    put:
      operationId: updateNotice
      tags: [notices]
      summary: Update notice (admin)
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    delete:
      operationId: deleteNotice
      tags: [notices]
      summary: Delete notice (admin)
      parameters:
        - name: noticeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        batchId:
          type: integer
        icon:
          type: string
          nullable: true
        order:
          type: integer
        lecturesCount:
          type: integer
        notesCount:
          type: integer
      required: [id, name, batchId, order, lecturesCount, notesCount]
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        batchId:
          type: integer
        icon:
          type: string
          nullable: true
        order:
          type: integer
      required: [name, batchId, order]
    SubjectListResponse:
      type: object
      properties:
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
      required: [subjects]

    # LECTURE SCHEMAS
...
    Question:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
          nullable: true
        marks:
          type: number
      required: [id, text, optionA, optionB, optionC, optionD, correctOption, marks]
    QuestionInput:
      type: object
      properties:
        text:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
          nullable: true
        marks:
          type: number
      required: [text, optionA, optionB, optionC, optionD, correctOption, marks]
    Test:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [test, quiz]
        batchId:
          type: integer
        subjectId:
          type: integer
          nullable: true
        totalMarks:
          type: number
        durationMinutes:
          type: integer
        isActive:
          type: boolean
        questionsCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, type, batchId, totalMarks, durationMinutes, isActive, questionsCount, createdAt]
    TestDetail:
      allOf:
        - $ref: "#/components/schemas/Test"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    CreateTestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [test, quiz]
        batchId:
          type: integer
        subjectId:
          type: integer
          nullable: true
        durationMinutes:
          type: integer
        isActive:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
      required: [title, type, batchId, durationMinutes, isActive]
    SubmitTestBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              selectedOption:
                type: string
                enum: [A, B, C, D]
            required: [questionId, selectedOption]
      required: [answers]
    TestResult:
      type: object
      properties:
        id:
          type: integer
        testId:
          type: integer
        userId:
          type: integer
        score:
          type: number
        totalMarks:
          type: number
        percentage:
          type: number
        timeTaken:
          type: integer
          nullable: true
        submittedAt:
          type: string
      required: [id, testId, userId, score, totalMarks, percentage, submittedAt]
    TestListResponse:
      type: object
      properties:
        tests:
          type: array
          items:
            $ref: "#/components/schemas/Test"
      required: [tests]
    TestResultListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TestResult"
      required: [results]

    # NOTICE SCHEMAS
    Notice:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [notice, offer]
        batchId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, title, content, type, isActive, createdAt]
    CreateNoticeBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [notice, offer]
        batchId:
          type: integer
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
      required: [title, content, type, isActive]
    NoticeListResponse:
      type: object
      properties:
        notices:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
      required: [notices]

    # ADMIN SCHEMAS
...