# lib/api-zod/src/generated/types/submitTestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUPHY EduTech Platform API for Calcutta University Physics Honours Students
 * OpenAPI spec version: 1.0.0
 */
import type { SubmitTestBodyAnswersItem } from "./submitTestBodyAnswersItem";

export interface SubmitTestBody {
  answers: SubmitTestBodyAnswersItem[];
}